[FRC] Couple tokenizer with data packer (#311)
In the current code, `tokenizer` is instantiated and passed around in
multiple places. Notably, it's used in `DataPacker`, `Trainer`,
`vLLMRollout`, various subclasses of `Dataset`, `SFTTrainer`,
`GRPOTrainer`, `TRTLLMRollout`, `Controller` , various `StatusManager`
for policy/rollout, and in various custom SFT and GRPO implementations.
This proliferation of use has consequences:
1. It's impossible to use Cosmos-RL for any use-case that doesn't rely
on tokenizer for input processing (e.g., transformer-based image only
models such as DINO for object detection).
2. `tokenizer` is sometimes passed around just for convenience but not
used given the complexity of the rest of the code.
3. Separation of concerns is blurry between `tokenizer`, custom
`Dataset` and `DataPacker`.

To simplify the code-base as well as to make it more generalizable to
more use-cases, this PR removes the traces of `tokenizer` is most places
and **couples** the `tokenizer` with `DataPacker` implementation. After
this PR, the primary place that holds on to `tokenizer` is `DataPacker`.
Ironically, already all `DataPacker` implementations had the `tokenizer`
instance and used it.

In a follow-up PR, we'll introduce a new type of `DataPacker` that does
not rely on `tokenizer` for data reading and/or decoding, e.g. a
`DataPacker` that only reads images and transform/resize/etc images for
model to use.